+++ network/message/src/main/java/com/seugi/network/message/request/MessageRequest.kt
@@ -4,10 +4,10 @@ data class MessageRequest(
     val roomId: String,
     val type: String,
     val message: String,
-    val mention: List<Int>,
+    val mention: List<Long>,
     val mentionAll: Boolean,
     val emotion: String?,
     val uuid: String,
 ) {
-    constructor(roomId: String, message: String, uuid: String, type: String, mention: List<Int>) : this(roomId, type, message, mention, false, null, uuid)
+    constructor(roomId: String, message: String, uuid: String, type: String, mention: List<Long>) : this(roomId, type, message, mention, false, null, uuid)
 }
